Make VSGI more stubbable
Use 'GLib.IOStream' instead of having a proper class so that we can stub
with 'GLib.SimpleIOStream'.

Extract an 'AutoCloseIOStream' implementation from the SCGI server. This
can be used by implementation that expect the connection to be closed
when disposed.

Make both 'Request' and 'Response' derivable concrete classes and remove
'CGI' and 'Mock' implementations as this is now handled upstream.